feat: implement #81 - Decouple database scripts from application
