+++ .env.sample
@@ -7,20 +7,21 @@ E2B_API_KEY=
 GITHUB_PAT=
 CLOUDFLARED_TUNNEL_TOKEN=
 
-# Production Supabase (source for sync)
-SUPABASE_URL=https://gownillwfbtrbnkrvrxi.supabase.co
+# Supabase — environment-agnostic
+# The deployment environment determines which database these point to.
+# In local dev, set these to your staging instance.
+# In production, Vercel (or your hosting provider) sets these to the production instance.
+SUPABASE_URL=
 SUPABASE_KEY=
 SUPABASE_SERVICE_KEY=
 
-# Staging Supabase (destination for sync)
-# NOTE: Use SERVICE ROLE key (not anon key) for full sync functionality.
-# The service role key is required for:
-#   - Bucket creation and management
-#   - Table creation via exec_sql RPC (if configured)
-# The anon key will work for data sync only.
-SUPABASE_URL_STAGING=https://hdyqdnnwhvhmvsdqvkpu.supabase.co
-SUPABASE_KEY_STAGING=
-SUPABASE_SERVICE_KEY_STAGING=
+# Sync script (scripts/sync-supabase.ts) — CI/CD only
+# The staging env vars below are only needed when running the sync script directly.
+# They should be set in CI/CD secrets, not committed to .env.
+# SUPABASE_URL_STAGING=
+# SUPABASE_KEY_STAGING=
+# SUPABASE_SERVICE_KEY_STAGING=
+
 VERCEL_TOKEN=
 VERCEL_PROJECT_ID=prj_HQF1SOUMElzqB4p5W7MnfBfaWpnV
 VERCEL_ORG_ID=team_6hmRdTXofBIeNvqXxLB5EVQR
